add examples

+++ utils/format_image.py
@@ -2,4 +2,4 @@ def format_image(img):
     if img.max() <= 1:
         return (img * 255).astype('uint8')
     else:
-        return img
+        return img.astype('uint8')
